Update README.md
Made a few extra edits. Personally I think the WP repo slug should also be: colossus-wp-options-framework

The directory is a mess for searching, so keeping it the same as the plugin name would in my opinion be a good idea. But open to your suggestions.

I've also removed a few bits that aren't relevant.